try to add sqlite to lambda
