+++ packages/www/.gitignore
@@ -90,3 +90,4 @@ sw.*
 *.swp
 
 /*-dist/
+data.db
