+++ packages/www/package.json
@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "concurrently 'wait-on tcp:$SERVER_PORT && yarn dev:www' 'yarn dev:server'",
-    "build": "yarn build:server && concurrently 'wait-on tcp:$SERVER_PORT && yarn build:www && kill-port $SERVER_PORT || :' 'yarn start:server || :'",
+    "dev": "concurrently 'yarn dev:lambda' 'wait-on tcp:$SERVER_PORT && yarn dev:www' 'yarn dev:server'",
+    "build": "yarn build:lambda && yarn build:server && concurrently 'wait-on tcp:$SERVER_PORT && yarn build:www && kill-port $SERVER_PORT || :' 'yarn start:server || :'",
     "start": "concurrently 'wait-on tcp:$SERVER_PORT && yarn start:www' 'yarn start:server'",
     "dev:www": "nuxt",
     "build:www": "nuxt generate",
@@ -14,7 +14,11 @@
     "dev:server": "NODE_ENV=development ts-node-dev -P server/tsconfig.json -r tsconfig-paths/register server/index.ts",
     "build:server": "tsc -P server/tsconfig.json",
     "start:server": "node -r tsconfig-paths/register server-dist/index.js",
-    "ts": "ts-node -P server/tsconfig.json -r tsconfig-paths/register"
+    "ts": "ts-node -P server/tsconfig.json -r tsconfig-paths/register",
+    "make-db": "yarn ts scripts/upload.ts",
+    "dev:lambda": "concurrently 'netlify-lambda serve lambda' 'wait-on ./lambda-dist && cp data.db lambda-dist/search/data.db'",
+    "build:lambda": "netlify-lambda build lambda && cp data.db lambda-dist/search/data.db",
+    "postinstall": "netlify-lambda install"
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.34",
@@ -22,6 +26,7 @@
     "@fortawesome/free-solid-svg-icons": "^5.15.2",
     "@nuxt/typescript-runtime": "^2.0.1",
     "@nuxtjs/fontawesome": "^1.1.2",
+    "@types/aws-lambda": "^8.10.72",
     "axios": "^0.21.1",
     "better-sqlite3": "^7.1.2",
     "concurrently": "^6.0.0",
@@ -30,6 +35,7 @@
     "fastify": "^3.14.0",
     "fastify-swagger": "^4.4.1",
     "jsonschema-definer": "^1.3.2",
+    "netlify-lambda": "^2.0.3",
     "nuxt": "^2.14.12",
     "nuxt-buefy": "^0.4.4",
     "nuxt-mq": "^2.0.2",
@@ -39,10 +45,15 @@
     "wait-on": "^5.2.1"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
+    "@babel/plugin-transform-object-assign": "^7.12.1",
+    "@babel/preset-typescript": "^7.12.1",
     "@nuxt/types": "^2.14.12",
     "@nuxt/typescript-build": "^2.0.4",
     "@nuxtjs/eslint-config-typescript": "^5.0.0",
     "@nuxtjs/eslint-module": "^3.0.2",
+    "@nuxtjs/proxy": "^2.1.0",
     "@patarapolw/scope-css": "^1.3.0",
     "@types/better-sqlite3": "^5.4.1",
     "@types/cheerio": "^0.22.27",
