+++ packages/www/server/api/index.ts
@@ -1,13 +1,11 @@
 /* eslint-disable no-throw-literal */
 /* eslint-disable require-await */
-import dayjs from 'dayjs'
 import { FastifyPluginAsync } from 'fastify'
 import swagger from 'fastify-swagger'
 import S from 'jsonschema-definer'
 
-import { ITheme, SEPARATOR, sTheme } from '../db'
+import { ITheme, SEPARATOR, parseQ, sTheme } from '../db'
 import { db } from '../db/sqlite'
-import { ISplitOpToken, splitOp } from '../db/tokenize'
 
 const apiRouter: FastifyPluginAsync = async (f) => {
   if (process.env.NODE_ENV === 'development') {
@@ -187,161 +185,13 @@ const apiRouter: FastifyPluginAsync = async (f) => {
         },
       },
       async (req): Promise<typeof sResponse.type> => {
-        const { page, limit } = req.query
-        let { q = '' } = req.query
-        q = q.trim()
+        const { page, limit, q } = req.query
 
-        let cond = ''
-
-        const params = {
-          map: new Map<number, string | number>(),
-          set(v: string | number) {
-            this.map.set(this.map.size, v)
-            return '$' + (this.map.size - 1)
-          },
-        }
-
-        if (q) {
-          const $and: string[] = []
-          const $or: string[] = []
-          const $nor: string[] = []
-
-          const segParse = (t: ISplitOpToken): string => {
-            switch (t.k) {
-              case 'tag':
-                return /* sql */ `"entry" MATCH ${params.set(
-                  `tag:"${t.v.replace(/"/g, '""')}"`
-                )}`
-              case 'title':
-                return `"entry" MATCH ${params.set(
-                  `title:"${t.v.replace(/"/g, '""')}"`
-                )}`
-              case 'path':
-                return `"entry" MATCH ${params.set(
-                  `path:"${t.v.replace(/"/g, '""')}"`
-                )}`
-              case 'content':
-                return `"entry" MATCH ${params.set(
-                  `text:"${t.v.replace(/"/g, '""')}"`
-                )}`
-              case 'date':
-                // eslint-disable-next-line no-case-declarations
-                let date: Date | null = null
-
-                if (/^\d+(\.\d+)?$/.test(t.v) || /^\.\d+/.test(t.v)) {
-                  date = dayjs()
-                    .subtract(Math.abs(parseFloat(t.v)), 'days')
-                    .toDate()
-                } else {
-                  let d: dayjs.Dayjs
-
-                  const m = /^(\d+(?:\.\d+)?)?([A-Z]{1,8})$/.exec(t.v)
-                  if (m) {
-                    d = dayjs().subtract(
-                      parseFloat(m[1]!),
-                      m[2] as dayjs.OpUnitType
-                    )
-                  } else {
-                    d = dayjs(t.v)
-                  }
-
-                  if (d.isValid()) {
-                    date = d.toDate()
-                  }
-                }
-
-                if (!date) {
-                  return /* sql */ `FALSE`
-                }
-
-                switch (t.op) {
-                  case '>':
-                    return /* sql */ `"date" > ${params.set(+date)}`
-                  case '<':
-                    return /* sql */ `"date" < ${params.set(+date)}`
-                }
-
-                return /* sql */ `"date" > ${params.set(+date)}`
-            }
-
-            return `"entry" MATCH ${params.set(
-              `(tag:"${t.v.replace(/"/g, '""')}" OR text:"${t.v.replace(
-                /"/g,
-                '""'
-              )}" OR title:"${t.v.replace(/"/g, '""')}")`
-            )}`
-          }
-
-          // eslint-disable-next-line array-callback-return
-          splitOp(q).map((t) => {
-            switch (t.prefix) {
-              case '-':
-                $nor.push(segParse(t))
-                break
-              case '?':
-                $or.push(segParse(t))
-                break
-              default:
-                $and.push(segParse(t))
-            }
-          })
-
-          if ($or.length) {
-            $and.push(`(${$or.join(' OR ')})`)
-          }
-
-          if ($nor.length) {
-            $and.push(`NOT (${$or.join(' AND ')})`)
-          }
-
-          cond = $and.length ? $and.join(' AND ') : 'FALSE'
-          cond = '"date" IS NOT NULL AND ' + cond
-        }
-
-        if (!cond) {
-          cond = '"date" IS NOT NULL'
-        }
-
-        const rCount = db
-          .prepare(
-            /* sql */ `
-        SELECT COUNT("path") AS count
-        FROM "entry"
-        WHERE ${cond}
-        `
-          )
-          .get({ ...params.map })
-
-        if (!rCount) {
-          return {
-            result: [],
-            count: 0,
-          }
-        }
-
-        const rResult = db
-          .prepare(
-            /* sql */ `
-        SELECT "path", title, "date", "image", tag, html
-        FROM "entry"
-        WHERE ${cond}
-        ORDER BY "date" DESC
-        LIMIT ${limit} OFFSET ${(page - 1) * limit}
-        `
-          )
-          .all({ ...params.map })
-
-        return {
-          result: rResult.map((entry) => ({
-            path: entry.path,
-            title: entry.title,
-            date: entry.date ? new Date(entry.date).toISOString() : undefined,
-            image: entry.image || undefined,
-            tag: entry.tag.split(/ /g).filter((el: string) => el),
-            html: entry.html.split(SEPARATOR)[0],
-          })),
-          count: rCount.count,
-        }
+        return parseQ(db, {
+          q,
+          limit,
+          page,
+        })
       }
     )
   }
